requirements.txt
Flask
PyPDF2
sentence-transformers
openai
python-dotenv
python-docx
markdown
beautifulsoup4
requests
uvicorn==0.30.1
python-multipart==0.0.9
jinja2==3.1.4